# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fenix API",
        "version": "0.1"
    },
    "paths": {
        "/admin/subscriptions": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get subscription transactions",
                "operationId": "6dedcc1980cb3f93833a089102919065",
                "responses": {
                    "200": {
                        "description": "List of subscription transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "device_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "product_id": {
                                                "type": "string",
                                                "example": "prod_123"
                                            },
                                            "receipt_token": {
                                                "type": "string",
                                                "example": "some-long-receipt-token"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2023-01-01 00:00:00"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "example": "2023-01-01 00:00:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}